New reusable method for getting PuppeteerOptions and Add defaultMinToWaitForLWC so it can be controlled for Large OS
