+++ lib/utilityservice.js
@@ -1474,3 +1474,68 @@ UtilityService.prototype.moveFolders = async function (source, target) {
     await fs.copy(source, target);
     await fs.removeSync(source);
 }
+
+UtilityService.prototype.getPuppeteerOptions = async function (jobInfo) {
+    let puppeteerOptions = { 
+        headless: true,
+        args: ['--no-sandbox']
+    };
+
+    let macChrome = path.join('/', 'Applications', 'Google Chrome.app', 'Contents', 'MacOS', 'Google Chrome');
+    let winChrome = path.join('/', 'Program Files', 'Google', 'Chrome', 'Application', 'chrome.exe');
+    let winChrome86 = path.join('/', 'Program Files (x86)', 'Google', 'Chrome', 'Application', 'chrome.exe');
+
+    if (jobInfo.puppeteerExecutablePath) {
+        puppeteerOptions.executablePath = jobInfo.puppeteerExecutablePath;
+    } else if (fs.existsSync(macChrome)) {
+        puppeteerOptions.executablePath = macChrome;
+    } else if (fs.existsSync(winChrome)) {
+        puppeteerOptions.executablePath = winChrome;
+    } else if (fs.existsSync(winChrome86)) {
+        puppeteerOptions.executablePath = winChrome86;
+    } else {
+        let chromiumDirLocal = path.join('.', 'node_modules', 'puppeteer', '.local-chromium');
+        if (fs.existsSync(chromiumDirLocal)) {
+            fs.readdirSync(chromiumDirLocal).forEach((file) => {
+                let macApp = path.join(chromiumDirLocal, file, 'chrome-mac', 'Chromium.app', 'Contents', 'MacOS', 'Chromium');
+                let linuxApp =  path.join(chromiumDirLocal, file, 'chrome-linux', 'chrome');
+                let winApp =  path.join(chromiumDirLocal, file, 'chrome-win', 'chrome.exe');
+
+                if (fs.existsSync(macApp)) {
+                    puppeteerOptions.executablePath = macApp;
+                } else if (fs.existsSync(linuxApp)) {
+                    puppeteerOptions.executablePath = linuxApp;
+                } else if (fs.existsSync(winApp)) {
+                    puppeteerOptions.executablePath = winApp;
+                }
+            });
+        }
+        if (!puppeteerOptions.executablePath) {
+            let pathToPuppeteer = require("global-modules-path").getPath("puppeteer", "puppeteer");
+            if (pathToPuppeteer) {
+                let chromiumDirGlobal = path.join(pathToPuppeteer, '.local-chromium');
+                if (fs.existsSync(chromiumDirGlobal)) {
+                    fs.readdirSync(chromiumDirGlobal).forEach((file) => {
+                        let macApp = path.join(chromiumDirGlobal, file, 'chrome-mac', 'Chromium.app', 'Contents', 'MacOS', 'Chromium');
+                        let linuxApp =  path.join(chromiumDirGlobal, file, 'chrome-linux', 'chrome');
+                        let winApp =  path.join(chromiumDirGlobal, file, 'chrome-win', 'chrome.exe');
+
+                        if (fs.existsSync(macApp)) {
+                            puppeteerOptions.executablePath = macApp;
+                        } else if (fs.existsSync(linuxApp)) {
+                            puppeteerOptions.executablePath = linuxApp;
+                        } else if (fs.existsSync(winApp)) {
+                            puppeteerOptions.executablePath = winApp;
+                        }
+                    });
+                }   
+            }
+        }
+    }
+
+    if (puppeteerOptions.executablePath) {
+        jobInfo.puppeteerExecutablePath = puppeteerOptions.executablePath;
+    }
+    //console.log(puppeteerOptions);
+    return puppeteerOptions;
+}
